requirements.txt
discord.py
pendulum
asyncpg
